Added checking for file already exists to link.py

+++ link.py
@@ -1,5 +1,6 @@
 #!/usr/bin/python
 import commands
+import os
 
 here = "/home/garoth/Configs/"
 
@@ -9,5 +10,7 @@ rules = [
 
 for rule in rules:
         rule[0] = here + rule[0]
-        print "Linking '" + rule[0] + "' to '" + rule[1] + "'"
-        commands.getstatusoutput("ln -s " + rule[0] + " " + rule[1])
+        if not os.path.exists(rule[1]):
+                print "Linking '" + rule[0] + "' to '" + rule[1] + "'"
+                commands.getstatusoutput("ln -s " + rule[0] + " " + rule[1])
+        print "Already Exists: '" + rule[1] + "' -- not linking."
